+++ app-frontend/src/app/services/projects/project.service.js
@@ -353,7 +353,8 @@ export default (app) => {
 
         addScenesToLayer(projectId, layerId, sceneIds) {
             return this.Project.addScenesToLayer(
-                {projectId, layerId}, sceneIds
+                {projectId, layerId},
+                sceneIds
             ).$promise;
         }
 
@@ -637,7 +638,7 @@ export default (app) => {
         }
 
         getProjectPermissions(project, user) {
-            //TODO replace uses with permissionsService.getEditableObjectPermission
+            // TODO replace uses with permissionsService.getEditableObjectPermission
             return this.$q((resolve, reject) => {
                 if (project.owner.id === user.id || project.owner === user.id) {
                     resolve([